+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2019-09-01
+
+### Added
+- New Renderer::___renderResultsJSON() method for rendering search results as a JSON string.
+- Additional CSS rules to make sure that visited links appear correctly in the default output.
+
 ## [0.9.1] - 2019-08-29
 
 ### Fixed